+++ device-minion-runtime/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-18427] Fixed the stats worker to handle trace disabled case and fixed the null pointer evict issue.
 - [SW-18522] Fixed minor build issues seen with clang
 - [SW-18522] More minor fixes for clang builds
+- [SW-18881] Fixing build warning.
 ### Security
 
 ## [0.19.0] - 2023-08-03